+++ test/services/BlockSigner.test.ts
@@ -32,8 +32,8 @@ describe('BlockSigner', () => {
     mockedChainContract = sinon.createStubInstance(ChainContract);
     mockedFeedProcessor = sinon.createStubInstance(FeedProcessor);
     settings = {
-      feedsFile: 'src/config/feeds.yaml',
-      feedsOnChain: 'src/config/feedsOnChain.yaml',
+      feedsFile: 'test/feeds/feeds.yaml',
+      feedsOnChain: 'test/feeds/feedsOnChain.yaml',
     } as Settings;
 
     container.bind('Logger').toConstantValue(mockedLogger);